requirements.txt
azure-functions
azure-identity
azure-synapse-artifacts
azure-mgmt-synapse
pyodbc